♻️ Refactor: 팔로우 페이지 구조변경 - 팔로우가 없을때를 따로 뺌
