+++ src/pages/FollowersPage.jsx
@@ -4,9 +4,9 @@ import styled from 'styled-components';
 import { useLocation } from 'react-router-dom';
 import UserFollowListItem from '../components/Follow/UserFollowListItem';
 import Header from '../layouts/Header/Header';
-import logoImg from '../assets/images/logo.png';
 import Loading from '../layouts/Loading/Loading';
 import { followerAPI } from '../api/apis/follow';
+import FollowNone from '../components/Follow/FollowNone';
 
 const Main = styled.main`
   width: 100%;
@@ -23,31 +23,11 @@ const Main = styled.main`
   }
 `;
 
-const FollowingNoneWrapper = styled.section`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: calc(100vh - 150px);
-  gap: 25px;
-`;
-
-const LogoImg = styled.img`
-  width: 150px;
-  filter: grayscale(90%);
-`;
-
-const FollowInfo = styled.h3`
-  font-size: 14px;
-  color: var(--sub-text-color);
-`;
-
 export default function FollowersPage() {
   const [isLoading, setIsLoading] = useState(true);
   const location = useLocation();
   const accountname = location.state?.accountname;
   const [userData, setUserData] = useState([]);
-  const loginAccountname = localStorage.getItem('accountname');
 
   useEffect(() => {
     const fetchPosts = async () => {
@@ -76,21 +56,7 @@ export default function FollowersPage() {
               </ul>
             </section>
           ) : (
-            <FollowingNoneWrapper>
-              {loginAccountname === accountname ? (
-                <>
-                  <LogoImg src={logoImg} alt="포칼 로고" />
-                  <FollowInfo>나를 팔로우하는 유저가 없습니다</FollowInfo>
-                </>
-              ) : (
-                <>
-                  <LogoImg src={logoImg} alt="포칼 로고" />
-                  <FollowInfo>
-                    @{accountname}님을 팔로우하는 유저가 없습니다
-                  </FollowInfo>
-                </>
-              )}
-            </FollowingNoneWrapper>
+            <FollowNone type="follower" accountname={accountname} />
           )}
         </Main>
       )}
